Bug 481994 - Some XxxFeatureForPattern classes call ICustomUndoablePattern#redo instead of ICustomUndoRedoPattern#postRedo
Cleaned-up, so that all methods (canUndo/Redo, preUndo/Redo, undo/redo,
postUndo/Redo) try to delegate to the pattern methods in case the
pattern implement ICustonUndoablePattern or ICustomUndoRedoPattern for
all methods implemented by the interfaces. In case a method is part of
both interfaces first ICustomUndoablePattern is checked for, because
that is the older interface (to keep compatibility).
This change might modify the behavior for certain tools that rely on the
old, not really consistent behavior.

Change-Id: Ic3ffed5ea6549ca272f043e7eaf65b0db388721f